refactor: extract chunk_store_commit.c
Move commit/rollback path (csCommitToMemory/File, chunkStoreCommit,
chunkStoreRollback, and commit-only helpers) out of chunk_store.c.
Promote open/sync/header/manifest helpers via chunk_store_int.h for
cross-TU use. Wire build, amalgamation, and lint.
